--- core/core-domain/src/main/java/com/pgms/coredomain/domain/common/event/EventReview.java
@@ -1,18 +1,0 @@
-package com.pgms.coredomain.domain.common.event;
-
-import com.pgms.coredomain.domain.common.BaseEntity;
-import jakarta.persistence.*;
-
-@Entity
-@Table(name = "event_review")
-public class EventReview extends BaseEntity {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    private Integer score;
-
-    @Lob
-    private String content;
-}
