--- core/core-domain/src/main/java/com/pgms/coredomain/domain/common/event/EventStatus.java
@@ -1,4 +1,0 @@
-package com.pgms.coredomain.domain.common.event;
-
-public enum EventStatus {
-}
