+++ core/core-domain/src/main/java/com/pgms/coredomain/domain/event/EventHall.java
@@ -1,4 +1,4 @@
-package com.pgms.coredomain.domain.common.event;
+package com.pgms.coredomain.domain.event;
 
 import com.pgms.coredomain.domain.common.BaseEntity;
 import jakarta.persistence.*;
