+++ core/core-domain/src/main/java/com/pgms/coredomain/domain/event/EventHallSeat.java
@@ -1,0 +1,16 @@
+package com.pgms.coredomain.domain.event;
+
+import com.pgms.coredomain.domain.common.BaseEntity;
+import jakarta.persistence.*;
+
+@Entity
+@Table(name = "event_hall_seat")
+public class EventHallSeat extends BaseEntity {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    private String name;
+
+}
