+++ core/core-domain/src/main/java/com/pgms/coredomain/domain/event/EventReview.java
@@ -1,16 +1,18 @@
-package com.pgms.coredomain.domain.common.event;
+package com.pgms.coredomain.domain.event;
 
 import com.pgms.coredomain.domain.common.BaseEntity;
 import jakarta.persistence.*;
 
 @Entity
-@Table(name = "event_hall_seat")
-public class EventHallSeat extends BaseEntity {
+@Table(name = "event_review")
+public class EventReview extends BaseEntity {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
-    private String name;
+    private Integer score;
 
+    @Lob
+    private String content;
 }
