+++ core/core-domain/src/main/java/com/pgms/coredomain/domain/event/EventStatus.java
@@ -1,0 +1,4 @@
+package com.pgms.coredomain.domain.event;
+
+public enum EventStatus {
+}
